pyproject.toml
[tool.poetry]
name = "encrypt-game-key"
version = "0.1.0"
description = ""
authors = ["Sascha Vis <saschavis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0.2"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
